A csomag fogadásában hibák lettek javítva.
- FIX: megfelelő indexen keresi a receive() a node ID-t
- FIX: megfelelő hosszúságú tömböt foglal le a package számára

- ISSUE: a válasznál a "van/nincs" és az "ennyi van" rendszer egyszerre van jelen, és egyik sem működik

+++ MSCPSlaveSerial.h
@@ -34,7 +34,7 @@ class MSCPSlaveSerial
 		byte			_manualReplyLength;
 		unsigned long	_manualReplyOriginalMsgID;
 		byte			_manualReplyState;
-		unsigned long	_replyNeededCount;
+		uint8_t			_replyNeededCount;
 
 		// kapcsolat specifikus
 		byte			_connectionState;
